requirements.txt
# Core workflow + evaluator dependencies
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
PyYAML>=6.0

# CLI / UX
typer>=0.9,<1
rich>=13.0

# Workflow orchestration
langgraph>=0.2,<1

# LLM + retry
litellm>=1,<2
tenacity>=8.0
httpx>=0.24

# Tool graph generation (s08)
networkx>=3.0

# Unit test (s09)
pytest